Switched threads to pthread for all platforms (on Windows, you can compile with Win32 threads using PTHREAD=n).
