+++ src/win/CMakeLists.txt
@@ -15,7 +15,7 @@
 
 enable_language(RC)
 
-add_library(plat OBJECT win.c win_dynld.c win_thread.c win_cdrom.c
+add_library(plat OBJECT win.c win_dynld.c win_cdrom.c
 	win_keyboard.c win_crashdump.c win_midi.c win_mouse.c)
 
 add_library(ui OBJECT win_ui.c win_stbar.c win_sdl.c win_dialog.c win_about.c
